Added support for Landmarks dataset.
